simply float type mapped with java double

+++ src/main/java/ast/util/enums/DataType.java
@@ -4,7 +4,7 @@ import com.fasterxml.jackson.annotation.JsonAlias;
 
 public enum  DataType {
     @JsonAlias({"int"}) IntegerType("integer"),
-    @JsonAlias({"float"}) FloatType("float"),
+    @JsonAlias({"double"}) FloatType("float"),
     @JsonAlias({"boolean"}) BooleanType("boolean"),
     @JsonAlias({"char"}) CharType("char"),
     @JsonAlias({"String", "string"}) StringType("string"),
